[MVVM] - Refactor pkgs and adde ViewModel
