+++ app/src/main/java/com/example/login_compose/login/LoginViewModel.kt
@@ -1,0 +1,6 @@
+package com.example.login_compose.login
+
+import androidx.lifecycle.ViewModel
+
+class LoginViewModel: ViewModel() {
+}
