+++ app/src/main/java/com/example/login_compose/login/login/HeaderLogin.kt
@@ -1,4 +1,4 @@
-package com.example.login_compose.components.login
+package com.example.login_compose.login.login
 
 import android.app.Activity
 import androidx.compose.foundation.clickable
